Moved exportutils to utils directory
